// src/api/models/updateMembershipRoleBodyRole.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

export type UpdateMembershipRoleBodyRole = typeof UpdateMembershipRoleBodyRole[keyof typeof UpdateMembershipRoleBodyRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateMembershipRoleBodyRole = {
  COMMUNITY_MEMBER: 'COMMUNITY_MEMBER',
  COMMUNITY_ADMIN: 'COMMUNITY_ADMIN',
} as const;